src/main/java/practice/leecode/String/StrStr_28.java
package practice.leecode.String;

public class StrStr_28 {

    public int strStr(String haystack, String needle) {
        if (needle.length() == 0) {
            return 0;
        }
        for (int i = 0, j = 0; i < haystack.length() && j < needle.length();) {
            if (haystack.charAt(i) == needle.charAt(j)) {
                j++;
                i++;
            } else {
                i -= j - 1;
                j = 0;
            }
            if (j == needle.length()) {
                return i - j;
            }
        }
        return -1;
    }
}
